Widen object and linker state to 64 bits

+++ include/linker.h
@@ -30,7 +30,7 @@ typedef struct {
     bool verbose;
     bool shared;        /* Create shared library (.rll) */
     bool shared_explicit;
-    uint32_t base_addr; /* Base load address */
+    uint64_t base_addr; /* Base load address */
     const char* entry;  /* Entry point symbol */
     bool entry_explicit;
     char manifest_entry[RCC_MAX_IDENT];
@@ -50,18 +50,18 @@ typedef struct LinkedSection {
     SectionType type;
     uint32_t flags;
     uint8_t* data;
-    uint32_t size;
-    uint32_t capacity;
+    uint64_t size;
+    uint64_t capacity;
     uint32_t align;
-    uint32_t vaddr;     /* Virtual address after linking */
+    uint64_t vaddr;     /* Virtual address after linking */
     struct LinkedSection* next;
 } LinkedSection;
 
 /* Global symbol (after linking) */
 typedef struct GlobalSymbol {
     const char* name;
-    uint32_t value;     /* Final resolved address */
-    uint32_t size;
+    uint64_t value;     /* Final resolved address */
+    uint64_t size;
     SymbolType type;
     SymbolBinding binding;
     int section;        /* Index into linked sections */
@@ -72,10 +72,10 @@ typedef struct GlobalSymbol {
 
 /* Pending relocation */
 typedef struct PendingReloc {
-    uint32_t offset;        /* Offset in linked section */
+    uint64_t offset;        /* Offset in linked section */
     const char* symbol;     /* Symbol name */
     RelocType type;
-    int32_t addend;
+    int64_t addend;
     int section;            /* Which linked section */
     const char* source;     /* Source object file */
     struct PendingReloc* next;
@@ -100,8 +100,8 @@ typedef struct Linker {
     int reloc_count;
 
     /* Layout info */
-    uint32_t base_addr;
-    uint32_t entry_addr;
+    uint64_t base_addr;
+    uint64_t entry_addr;
 } Linker;
 
 /* Linker functions */
@@ -117,7 +117,7 @@ bool linker_merge_sections(Linker* ld);
 bool linker_collect_symbols(Linker* ld);
 bool linker_resolve_symbols(Linker* ld);
 bool linker_apply_relocations(Linker* ld);
-bool linker_layout(Linker* ld, uint32_t base_addr);
+bool linker_layout(Linker* ld, uint64_t base_addr);
 
 /* Output */
 bool linker_emit_rin(Linker* ld, const char* filename);
